Refactored how vertex layouts are managed. The old code was overengineered and may have been causing unnecessary PSO duplicates Moved vertexLayoutId from the Hlms hash and moved into a Renderable property. This allows us to support many more layouts (up to 1024 without operation types, from 255 including operation types) VertexLayout IDs between v1 & v2 are now shared, which reduces shader & PSO unnecessary duplication VertexLayouts now live throughout the lifetime of the application (which shouldn't be a problem given how many we can create) which is necessary for HlmsDiskCache to work correctly

+++ OgreMain/include/Vao/OgreVertexArrayObject.h
@@ -78,9 +78,9 @@ namespace Ogre
         ///     In OpenGL mInputLayoutId is always 0. mVaoName changes when either the
         ///     layout or the internal vertex/index buffer changes
         ///
-        ///     In the rest of the systems same vertex layouts share the same ID,
+        ///     In the rest of the systems, same vertex layouts share the same ID,
         ///     mVaoName changes when the vertex/index buffer needs to change.
-        uint8 mInputLayoutId;
+        uint16 mInputLayoutId;
 
         uint32                  mPrimStart;
         uint32                  mPrimCount;
@@ -100,7 +100,7 @@ namespace Ogre
 
         uint32 getRenderQueueId(void) const                             { return mRenderQueueId; }
         uint32 getVaoName(void) const                                   { return mVaoName; }
-        uint8 getInputLayoutId(void) const                              { return mInputLayoutId; }
+        uint16 getInputLayoutId(void) const                             { return mInputLayoutId; }
 
         const VertexBufferPackedVec& getVertexBuffers(void) const       { return mVertexBuffers; }
         IndexBufferPacked* getIndexBuffer(void) const                   { return mIndexBuffer; }
